git-svn-id: file:///home/gecchi/repos@1361 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/src/jp/ggaf/dx9core/util/GgafDx9GeometryMorpher.cpp
@@ -108,14 +108,14 @@ void GgafDx9GeometryMorpher::intoTargetLinerStep(int prm_target_mesh, float prm_
 void GgafDx9GeometryMorpher::loopLiner(int prm_target_mesh, DWORD prm_loop_spend_frame, float prm_loop_num) {
     _method[prm_target_mesh] = LOOP_LINER;
     _halfloop_cnt[prm_target_mesh] = 0;
-    _stop_halfloop_num[prm_target_mesh] = prm_loop_num*2;
+    _stop_halfloop_num[prm_target_mesh] = (int)(prm_loop_num*2.0);
     _velo_weight[prm_target_mesh] = 1.0 / (prm_loop_spend_frame / 2);
 }
 
 void GgafDx9GeometryMorpher::loopTriangleWave(int prm_target_mesh, DWORD prm_loop_spend_frame, DWORD prm_attack_frame, DWORD prm_rest_frame, float prm_loop_num) {
     _method[prm_target_mesh] = LOOP_TRIANGLEWAVE;
     _halfloop_cnt[prm_target_mesh] = 0;
-    _stop_halfloop_num[prm_target_mesh] = prm_loop_num*2;
+    _stop_halfloop_num[prm_target_mesh] = (int)(prm_loop_num*2.0);
 
     _loop_attack_frame[prm_target_mesh] = prm_attack_frame;
     _loop_rest_frame[prm_target_mesh] = prm_rest_frame;
